move related code together
